Periodic merge

See merge request cmsinnertracker/Ph2_ACF!96

+++ src/CMSITminiDAQ.cc
@@ -75,7 +75,7 @@ void interruptHandler(int handler)
 void readBinaryData(const std::string& binaryFile, SystemController& mySysCntr, std::vector<RD53Event>& decodedEvents)
 {
     const unsigned int    wordDataSize = 32; // @CONST@
-    unsigned int          errors       = 0;
+    size_t                errors       = 0;
     std::vector<uint32_t> data;
 
     RD53Event::ForkDecodingThreads();
